Add initial DCS APIs
Add the gRPC dependencies and the gRPC generated files. We will want
to change to some other solution in future rather than committing the
generated files but this is good enough for the moment.

Implement the initial request/response API calls that are currently
implemented in the server @ https://github.com/DCS-gRPC/rust-server

This implementation in the server and client gets us to the
proof-of-concept stage.