Add DCS event stream processing
Add DCS event stream processing in a way that is easiest for processing.

We will convert the gRPC response objects to hashes to hide implementation
details while keeping things easy to process. We will flatten the event
hash and set the `:event` key to contain the event type.

This method will block until the streaming terminates  so should probably
be called in a thread for any application that is not only processing these
events but other things as well.